Allow assigning "starting inputs" with a scenario
When setting up a new scenario, instead of creating the scenario in a
before block, then assigning all of the starting input values
separately, like so...

  @scenario = Scenario.new(...)
  @scenario.input_1 = 1.0
  @scenario.input_2 = 10.5

...instead, assign them with the scenario...

  @scenario = Scenario.new(..., inputs: {
    input_1: 1.0, input_2: 10.5
  })

This will ensure that when comparing changes in the future or present
value, these "starting inputs" are used when getting the previous query
values.

Using the old method of assigning inputs separately will result in the
"previous" present and future values coming from a blank scenario.